Classic solution mess around - rename src to Graveyard - move Graveyard.csproj to Graveyard folder - Switch MahApps.Metro to nuget package reference - Switch HearthDb to local project reference - Change Debug/Release platform to x86 - Link plugin version to assembly version
