--- Graveyard.csproj
@@ -1,92 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">x86</Platform>
-    <ProjectGuid>{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>Graveyard</RootNamespace>
-    <AssemblyName>Graveyard</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="PresentationCore" />
-    <Reference Include="PresentationFramework" />
-    <Reference Include="System" />
-    <Reference Include="System.Xaml" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Drawing" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Windows.Forms" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Xml" />
-    <Reference Include="WindowsBase" />
-    <Reference Include="HearthDb">
-      <HintPath>lib\HearthDb.dll</HintPath>
-    </Reference>
-    <Reference Include="HearthstoneDeckTracker">
-      <HintPath>lib\HearthstoneDeckTracker.exe</HintPath>
-    </Reference>
-    <Reference Include="MahApps.Metro">
-      <SpecificVersion>False</SpecificVersion>
-      <HintPath>lib\MahApps.Metro.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="src\ChancesTracker.cs" />
-    <Compile Include="src\DragoncallerAlannaView.cs" />
-    <Compile Include="src\HadronoxView.cs" />
-    <Compile Include="src\GuldanView.cs" />
-    <Compile Include="src\Properties\AssemblyInfo.cs" />
-    <Compile Include="src\DiscardView.cs" />
-    <Compile Include="src\GraveyardPlugin.cs" />
-    <Compile Include="src\AnyfinCalculator.cs" />
-    <Compile Include="src\AnyfinView.cs" />
-    <Compile Include="src\NormalView.cs" />
-    <Compile Include="src\ResurrectView.cs" />
-    <Compile Include="src\NZothView.cs" />
-    <Compile Include="src\InputManager.cs" />
-    <Compile Include="src\Graveyard.cs" />
-    <Compile Include="src\ShudderwockView.cs" />
-    <Compile Include="src\SettingsView.xaml.cs">
-      <DependentUpon>SettingsView.xaml</DependentUpon>
-    </Compile>
-    <Page Include="src\SettingsView.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Compile Include="src\Settings.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTimeSharedInput>True</DesignTimeSharedInput>
-      <DependentUpon>Settings.settings</DependentUpon>
-    </Compile>
-    <None Include="src\Settings.settings">
-      <Generator>PublicSettingsSingleFileGenerator</Generator>
-      <LastGenOutput>Settings.Designer.cs</LastGenOutput>
-    </None>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Target Name="AfterBuild" Condition=" '$(Configuration)' == 'Debug' ">
-    <Copy SourceFiles="bin/Debug/Graveyard.dll" DestinationFolder="$(AppData)\HearthstoneDeckTracker\Plugins" />
-  </Target>
-</Project>
