+++ Graveyard.sln
@@ -1,22 +1,41 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28803.156
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Graveyard", "Graveyard.csproj", "{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Graveyard", "Graveyard\Graveyard.csproj", "{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HearthDb", "HearthDb\HearthDb\HearthDb.csproj", "{7ED14243-E02B-4B94-AF00-A67A62C282F0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Debug|x86.ActiveCfg = Debug|x86
+		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Debug|x86.Build.0 = Debug|x86
 		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Release|x86.ActiveCfg = Release|x86
+		{59C8CCA4-7F9B-4592-87A8-A23ADED7759E}.Release|x86.Build.0 = Release|x86
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Debug|x86.ActiveCfg = Debug|x86
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Debug|x86.Build.0 = Debug|x86
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Release|x86.ActiveCfg = Release|x86
+		{7ED14243-E02B-4B94-AF00-A67A62C282F0}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A5703983-F932-4DBC-94F0-0FE31B4E6A90}
 	EndGlobalSection
 EndGlobal
